requirements.txt
# AI Framework (--pre required during preview)
agent-framework-azure-ai>=0.1.0.dev0

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
# psycopg2-binary==2.9.9  # PostgreSQL adapter (optional for production)
aiosqlite==0.19.0       # SQLite async support

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.1

# Data Processing  
pydantic>=2.9.0
python-dateutil==2.8.2

# File Generation
markdown==3.5.1
weasyprint==60.2        # HTML to PDF
jinja2==3.1.2

# CLI Interface
click==8.1.7

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0

# Optional: Vector Database for advanced matching
# chromadb==0.4.18